+++ contracts/pools/IDonor.sol
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.6.12;
+
+// @todo find better name
+interface IDonor {
+    
+    function donate(uint256 _tokenId) external;
+}
