+++ contracts/tokens/IToken.sol
@@ -1,0 +1,18 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.6.12;
+
+// @todo interface to call erc1155 token implementation
+interface IToken {
+    
+    function mintCampaign(
+        address _to,
+        uint256 _voteId,
+        bytes calldata /*_campaignHash*/
+    ) external;
+    
+    function getBeneficiary(uint256 _tokenId) external view returns(address);
+    
+    function exists(uint256 _tokenId) external view returns (bool);
+    
+}
